Introduce reducable demand
The reducable demand should be listed within sinks, and provided an
efficiency (number between 0 and 1). This efficiency correspond to the
percent of the demand which must be provided (critical demand). The
oemof-solph sinks which models the non-critical part of the demand has
very small variable_costs such that it should not influence the costs
calculations but should be fulfilled rather than dumping energy into
excess sinks.